+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## [v1.0.0] - 2025-03-07
+## [v0.2.0] - 2025-03-20
+### Added
+- `--version` flag to check the current CLI version
+- Runtime check for Docker daemon availability
+
+## [v0.1.0] - 2025-03-07
 ### Added
 - Initial release with core features:
   - build administrative and development Docker images